api_foto/requirements.txt
asgiref==3.4.1
Django==3.2.6
django-appconf==1.0.4
django-imagekit==4.0.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
pilkit==2.0
Pillow==8.3.1
PyJWT==2.1.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
